pyproject.toml
[build-system]
requires = ["setuptools", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "pyopmnearwell"
description = "A Python framework to simulate near well dynamics using OPM Flow"
authors= [{ name = "David Landa-Marbán", email = "dmar@norceresearch.no" },
          {name = "Peter Moritz von Schultzendorff", email = "peter.schultzendorff@uib.no"},
          {name = "Antoine Lechevallier", email = "anle@norceresearch.no"}]
maintainers = [{ name = "David Landa-Marbán", email = "dmar@norceresearch.no" },
          {name = "Peter Moritz von Schultzendorff", email = "peter.schultzendorff@uib.no"},
          {name = "Antoine Lechevallier", email = "anle@norceresearch.no"}]
dynamic = ["version"]
readme = {file = "README.md", content-type = "text/markdown"}
license = {text = "GPL-3.0"}
keywords = ["co2", "dynamics", "hydrogen", "opm", "wells", "salt-precipitation", "corner-point-grid"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Natural Language :: English",
    "Topic :: Scientific/Engineering",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent"
]
dependencies = [
    "resdata",
    "mako",
    "matplotlib",
    "pandas",
    "scipy",
    "tensorflow",
    "keras_tuner",
    "scikit-learn"
]
requires-python = ">=3.12"

[tool.setuptools.dynamic]
version = {attr = "pyopmnearwell.__version__"}

[project.scripts]
pyopmnearwell = "pyopmnearwell.core.pyopmnearwell:main"

[project.urls]
Homepage = "https://cssr-tools.github.io/pyopmnearwell"
Repository = "https://github.com/cssr-tools/pyopmnearwell"

[tool.setuptools]
packages.find.where = ["src"]
platforms = ["all"]

[project.optional-dependencies]
opm = ["opm"]